refactor(@Studio/hooks): Created a new library for hooks, @studio/hooks (#13356)
* added @studio/hooks library
