+++ frontend/dashboard/pages/Dashboard/Dashboard.tsx
@@ -3,8 +3,9 @@ import classes from './Dashboard.module.css';
 import cn from 'classnames';
 import type { ChangeEvent, KeyboardEvent } from 'react';
 import { Textfield } from '@digdir/designsystemet-react';
-import { StudioButton, useDebounce } from '@studio/components';
+import { StudioButton } from '@studio/components';
 import { XMarkIcon } from '@studio/icons';
+import { useDebounce } from '@studio/hooks';
 import { CenterContainer } from '../../components/CenterContainer';
 import { DataModelsReposList } from '../../components/DataModelsRepoList';
 import { OrgReposList } from '../../components/OrgRepoList';
