+++ frontend/libs/studio-components/src/hooks/index.ts
@@ -3,5 +3,3 @@ export * from './usePrevious';
 export * from './useRetainWhileLoading';
 export * from './useLocalStorage';
 export * from './webStorage';
-export * from './useDebounce';
-export * from './useUniqueKeys';
