+++ frontend/libs/studio-hooks/README.md
@@ -1,0 +1,5 @@
+# @studio/hooks
+
+`@studio/hooks` is a library designed for reusable and generic hooks that can be easily shared across projects. Hooks that
+are tightly coupled with a specific feature should not be included here. Instead, they should be placed as close as
+possible to the relevant source code.
