+++ frontend/libs/studio-hooks/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require('../../jest.config');
