+++ frontend/libs/studio-hooks/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@studio/hooks",
+  "packageManager": "yarn@4.4.0",
+  "version": "0.1.0",
+  "main": "./src/index.ts",
+  "scripts": {
+    "test": "jest",
+    "lint": "eslint \"src/**/*.ts*\""
+  },
+  "dependencies": {
+    "@studio/pure-functions": "workspace:^",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "uuid": "10.0.0"
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "6.4.8",
+    "@testing-library/react": "16.0.0",
+    "@types/jest": "^29.5.5",
+    "eslint": "8.57.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "typescript": "^5.2.2"
+  }
+}
