+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigContent/EditActions/ActionsEditor/CustomActions/CustomActions.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { StudioTextfield, useDebounce } from '@studio/components';
+import { StudioTextfield } from '@studio/components';
+import { useDebounce } from '@studio/hooks';
 import { HelpText, Switch } from '@digdir/designsystemet-react';
 import {
   BpmnActionModeler,
