+++ frontend/packages/ux-editor/src/components/Properties/PageConfigPanel/HiddenExpressionOnLayout.test.tsx
@@ -22,7 +22,7 @@ const defaultLayouts: IFormLayouts = {
   [layout1NameMock]: layoutMock,
 };
 
-jest.mock('@studio/components/src/hooks/useDebounce.ts', () => ({
+jest.mock('@studio/hooks/src/hooks/useDebounce.ts', () => ({
   useDebounce: jest.fn().mockReturnValue({
     debounce: jest.fn((fn) => fn()),
   }),
